+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 ## [2025.2.5.1]
 
 <cite>Release contributors</cite>
-- 5 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.2.5.1+author%3Amlytvyn+is%3Apr)
+- 6 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.2.5.1+author%3Amlytvyn+is%3Apr)
 
 ### `Project Import` enhancements
 - Automatically download 3rd-party library sources [#1669](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1669)
 - Automatically download 3rd-party library javadocs [#1670](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1670)
+- Rely on Maven repositories configured via `Remote Jar Repositories` to download library sources [#1671](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1671)
 
 ### `hAC` enhancements
 - Added collapsible groups for connection dialog to target low-resolution screens [#1664](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1664)